refactor: move Checkbox component to form directory and update import in CertificatesTableColumns
